docs: update readme to clarify checkout step and env_vars usage
It seems to not be uncommon for people to forget the checkout step when using this action, so now it is included in each of the code examples.

`env_vars` clears all existing variables regardless of whether it's passed or not. Simply making it _Optional_ doesn't necessarily highlight this, so it has been made more clear.