Update README.md
## Why
Because when using a JSON env var it fails with a message that says:

```
google-github-actions/deploy-cloud-functions failed with: failed to parse KEY=VALUE pair ""my_key":"my_value"": missing "="
```

This was happening because an env var was a JSON string and commas were not escaped. Therefore when parsing the env vars it failed when reaching the second key of the JSON string.

## What
Adds a clarification to the "env_vars" parameter so that the user knows they need to escape their commas. The clarification was borrowed from the [source code of the function that parses the env vars](https://github.com/google-github-actions/actions-utils/blob/main/src/kv.ts#L43)

Signed-off-by: Juan Manuel Ramallo <juanmanuelramallo@hey.com>