Switch to gen2 API (#409)
This PR moves everything to use the `v2` API and adopts the "GEN_2"
function type. This is a major breaking change that should be released
as v3. As per our support policy, we will support v2 and v3, with v2
receiving critical bug fixes. We will drop support for v1.

- Fixes
https://github.com/google-github-actions/deploy-cloud-functions/issues/304
- Closes
https://github.com/google-github-actions/deploy-cloud-functions/issues/384
(N/A in GEN_2)
- Fixes
https://github.com/google-github-actions/deploy-cloud-functions/issues/373
- Closes
https://github.com/google-github-actions/deploy-cloud-functions/issues/399
(Cloud Functions recommends against CI/CD systems controlling IAM
permissions)
- Closes
https://github.com/google-github-actions/deploy-cloud-functions/issues/9
(N/A)